services/ai-diagnosis-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.1
python-multipart==0.0.20
Pillow==11.1.0
numpy==2.2.1
torch==2.5.1
torchvision==0.20.1
timm==1.0.12
onnxruntime==1.20.1
redis==5.2.1
pymongo==4.10.1
gridfs
scikit-learn==1.6.1
